backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
anthropic==0.34.2
pdfplumber==0.11.4
sqlalchemy==2.0.35
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
